pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "podcast-nuggets"
version = "0.1.0"
description = "Extract valuable nuggets from podcasts, videos, and audio"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Pontus Skog"}
]
keywords = ["podcast", "transcription", "notes", "learning", "whisper"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "anthropic>=0.39.0",
    "python-dotenv>=1.0.0",
    "youtube-transcript-api>=0.6.0",  # YouTube captions (KRITISK)
    "yt-dlp>=2024.1.0",               # YouTube metadata + audio download
]

[project.optional-dependencies]
whisper = [
    "mlx-whisper>=0.1.0",             # Apple Silicon transcription (fallback)
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.scripts]
nuggets = "nuggets.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["nuggets"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "mlx_whisper.*",
    "yt_dlp.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
